maint: Refactor plotter module (#725)
