+++ src/ansys/geometry/core/plotting/__init__.py
@@ -21,5 +21,6 @@
 # SOFTWARE.
 """Provides the PyAnsys Geometry plotting subpackage."""
 
-from ansys.geometry.core.plotting.plotter import Plotter, PlotterHelper
+from ansys.geometry.core.plotting.plotter import Plotter
+from ansys.geometry.core.plotting.plotter_helper import PlotterHelper
 from ansys.geometry.core.plotting.trame_gui import _HAS_TRAME, TrameVisualizer
